+++ test-parts/part-h-api_requests/api-main.js
@@ -2,11 +2,9 @@ const cAdminFile = require("./requests/req-admin_main");
 const cHealthFile = require("./requests/req-admin_health");
 const cAlarmFile = require("./requests/req-alarm");
 const cStorageFile = require("./requests/req-storage");
-/*
 const cDevicesFileCrud = require("./requests/req-devices_crud");
-const cDevicesFileCrudInvalid = require("./requests/req-devices_crud_invalid");
-const cDevicesFileModifyInvalid = require("./requests/req-devices_modify_invalid");
-*/
+//const cDevicesFileCrudInvalid = require("./requests/req-devices_crud_invalid");
+//const cDevicesFileModifyInvalid = require("./requests/req-devices_modify_invalid");
 
 function coordinateApiRequests()
 {
@@ -16,11 +14,9 @@ function coordinateApiRequests()
 		cHealthFile.callTestHealthApi();
 		cAlarmFile.callTestAlarmApis();
 		cStorageFile.callTestStorageAPIs();
-		/*
 		cDevicesFileCrud.callTestDeviceCrudApis();
-		cDevicesFileCrudInvalid.callTestDeviceCrudInvalidApis();
-		cDevicesFileModifyInvalid.callTestDeviceModifyInvalidApis();
-		*/
+		//cDevicesFileCrudInvalid.callTestDeviceCrudInvalidApis();
+		//cDevicesFileModifyInvalid.callTestDeviceModifyInvalidApis();
 	});
 }